requirements.txt
# 核心深度学习框架
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Hugging Face 生态系统
transformers>=4.37.0
datasets>=2.15.0
accelerate>=0.25.0
huggingface-hub>=0.20.0
tokenizers>=0.15.0
safetensors>=0.4.0

# Qwen-VL 专用
qwen-vl-utils[decord]>=0.0.8

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0

# 数据处理和科学计算
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 工具和工具库
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
tensorboard>=2.14.0
requests>=2.31.0

# 其他必要依赖
protobuf>=4.24.0
filelock>=3.12.0
packaging>=23.0